Add styles to section 2 & 3 and fixed positioning.
Added styles to section 3 to match the sample.

Made changes to the .html and .css files to use
as code was previously using absolute values. This
caused the site to look weird when resizing
browser.

Utilized max-height and margin: auto
instead of absolute values.

(This commit is long as I forgot to commit after
halfway into section 3. Decided to fix positioning
while I was at it. Will remember to do more
frequent commits in future.)